Added BytePredicate
Inspired by FastUtil, JSky will add predicates of primitive types. I am aware that this doesn't provide a big performance difference as doing the same thing with Collections, but hey, it isn't hard to implement these and we all love to feel that we are doing things the most optimized way possible. Primitive types predicates are meant to reduce the amount of objects being created, not to be faster, as the performance difference, while present, it is negligible unless testing billions of times at a time.