Gradle structure setup for multi-module project
Currently, the build task doesn't build all modules, however, the structure itself is done. We now have a "base" module, this is JSky itself, and a "yaml" module, this will be the module that contains yaml files, obviously. This module is separated because we will need to shade SnakeYaml and that will make the jar size quite bigger, this way, developers can just decide to not include the yaml module if they don't want to use yaml files for a way smaller jar