Added the CLICommandPool
The pool allows commands to be executed on another thread, for example, the main thread. The scanner runs on its own thread and adds commands to the pool, the pool is then executed on the main thread (For example) so commands are actually run on the main thread. This is completely optional, a pool is not required for the CLICommandManager to work. If no pool is provided, commands will be executed immediately on the scanner thread