Made KeyPattern more strict (See desc)
This mostly corrects edge cases, like using only dashes as a key, empty string keys, keys that contain more than two consecutive dashes, or end / start with dashes. There are only two formats allowed:
- nodashes
- with-dashes

And that's all. This annotation makes it easier to use the key pattern without unexpected errors, as it will warn you about an invalid pattern being used (At least if you are using IntelliJ). By the way, the reason I went with dashes instead of underscores ('_') or dots ('.') is just to make this pattern more "unique". as underscores are already being used by enum names (Upper case, I know) and using dots would conflict with the DataMap class, that uses them to separate between DIFFERENT keys, so combining that with this pattern would look like:
- parent-key.child-key
And both the parent key and child key would be respecting the KeyPattern annotation, as they are both INDIVIDUAL keys, even if they are related as parent and child keys.