Added tag format (100% branch coverage)
Alright, so from now on this format will be the standard on all projects based on JSky. The tag format allows to easily create tags that can be then used to build customizable objects from a simple tag-based string. This is similar to the already existing color, format and event patterns on SkyUtils and the embed patterns that JDS provides. Both projects will soon adapt the new tag system so the syntax is standardized.  This system supports:
- Escape characters (\< and \>)
- Nested tags
- Invalid tag handling